Fix: Choice/gather start ignored after comment
Previous commit revealed a flaw in scanner.c: We _only_
check for a choice/gather start at the end of the previous
line (or the start of the file). I know this made sense to
me at the time (something about not being able to roll back
if we advance too far). But it turns out that we can simply
scan when we need to, and we don't need to lug all this
state around. Nice.

This obviates the "need" for the START_OF_FILE symobl.