Blocks start and end at column 0 again
… but simpler and more robust this time.

Now I remember why I did this weird thing of parsing the next
block on a newline: That way, the intervening empty lines
wouldn't be parsed and included in the syntax tree between
two blocks of the same depth. But since that broke anyway
when there was a comment in the way, it's good that we got
rid of it. The price is that now a block owns all the whitespace
up to its next sibling, which is a bit annoying, but not
too bad.