Proper conflict between string & words
Now *most* test cases run fine. Still some kinks to iron out.

Note that since we’ve added a lone backslash as a valid part of a
string/text, there would be a conflict between a choice start starts
with an eval and a choice condition. But removing the explicit separator
cirumvents that; a backslash is just text now, so that problem solves
itself.