Parens no longer have a precedence
This means that there are now proper GLR conflicts between words and
parens, finally!

This also means that our syntax now accepts things as text where the ink
compiler would reject it (although the *balanced* version would result
in text nonetheless). But that’s a bit of a red flag (indicating that
we’re doing things fundamentally different from the compiler), but it
might not be a problem in practice. Strict adherance to all compiler
corner cases isn’t an obvious benefit.