Fix: Unary operators now properly conflict with words
The solution was to put the precedence number around the *content* of
the expression, not the whole expression.

Needed to adapt the corresponding test, because it was checking for the
wrong thing syntactically (forgot to wrap it in `paragraph` and forgot a
`content`). But the spirit of the test was right 🎉.