Postfix no longer an expression
I had over-interpreted `++` and `--`. The compiler only accepts them as
a toplevel `~ foo++` statement and nothing else. Which is neat, because
it makes the expression syntax simpler.

Then again, we’ll keep accepting it in expressions, because some users
might expect it to work, and we can be a little more forgiving with the
highlighing.